backup/requirements.txt
Flask
flask-cors
pymongo
flask-bcrypt
python-dotenv
